refactor: extract POI fetching logic into a separate function and clean up UploadButton component
